@@ -2,9 +2,10 @@ import copy
 import json
 import logging
 import time
+from collections.abc import Collection
 from functools import wraps
 from json import JSONDecodeError
-from typing import Any, Callable, Collection, Dict, Optional, Tuple, Union
+from typing import Any, Callable, Optional, Union
 
 import backoff
 import requests
@@ -33,7 +34,7 @@ def _read_mfa_code_from_stdin() -> str:
     return inputted_mfa_code
 
 
-def _mask_senritive_value_for_dict(data: Dict[str, Any], keys: Collection[str]) -> Dict[str, Any]:
+def _mask_senritive_value_for_dict(data: dict[str, Any], keys: Collection[str]) -> dict[str, Any]:
     """
     dictに含まれているセンシティブな情報を"***"でマスクします。
 
@@ -163,7 +164,7 @@ def _create_request_body_for_logger(data: Any) -> Any:  # noqa: ANN401
     )
 
 
-def _create_query_params_for_logger(params: Dict[str, Any]) -> Dict[str, Any]:
+def _create_query_params_for_logger(params: dict[str, Any]) -> dict[str, Any]:
     """
     ログに出力するためのquery_paramsを生成する。
      * AWS関係のcredential情報をマスクする。
@@ -282,15 +283,15 @@ class AnnofabApi(AbstractAnnofabApi):
     # Private Method
     #########################################
     @staticmethod
-    def _encode_query_params(query_params: Dict[str, Any]) -> Dict[str, Any]:
+    def _encode_query_params(query_params: dict[str, Any]) -> dict[str, Any]:
         """query_paramsのvalueがlist or dictのときは、JSON形式の文字列に変換する。
         `getAnnotationList` webapiで指定できる `query`などのように、2階層のquery_paramsに対応させる。
 
         Args:
-            query_params (Dict[str,Any]): [description]
+            query_params (dict[str,Any]): [description]
 
         Returns:
-            Dict[str, str]: [description]
+            dict[str, str]: [description]
         """
         new_params = {}
         if query_params is not None:
@@ -303,10 +304,10 @@ class AnnofabApi(AbstractAnnofabApi):
 
     def _create_kwargs(
         self,
-        params: Optional[Dict[str, Any]] = None,
-        headers: Optional[Dict[str, Any]] = None,
+        params: Optional[dict[str, Any]] = None,
+        headers: Optional[dict[str, Any]] = None,
         request_body: Optional[Any] = None,  # noqa: ANN401
-    ) -> Dict[str, Any]:
+    ) -> dict[str, Any]:
         """
         requestsモジュールのget,...メソッドに渡すkwargsを生成する。
 
@@ -326,7 +327,7 @@ class AnnofabApi(AbstractAnnofabApi):
                 else:
                     new_params[key] = value
 
-        kwargs: Dict[str, Any] = {
+        kwargs: dict[str, Any] = {
             "params": new_params,
             "headers": headers,
         }
@@ -381,10 +382,10 @@ class AnnofabApi(AbstractAnnofabApi):
         http_method: str,
         url: str,
         *,
-        params: Optional[Dict[str, Any]] = None,
+        params: Optional[dict[str, Any]] = None,
         data: Optional[Any] = None,  # noqa: ANN401
         json: Optional[Any] = None,  # pylint: disable=redefined-outer-name  # noqa: ANN401
-        headers: Optional[Dict[str, Any]] = None,
+        headers: Optional[dict[str, Any]] = None,
         stream: bool = False,
         raise_for_status: bool = True,
         **kwargs,
@@ -467,11 +468,11 @@ class AnnofabApi(AbstractAnnofabApi):
         http_method: str,
         url_path: str,
         *,
-        query_params: Optional[Dict[str, Any]] = None,
-        header_params: Optional[Dict[str, Any]] = None,
+        query_params: Optional[dict[str, Any]] = None,
+        header_params: Optional[dict[str, Any]] = None,
         request_body: Optional[Any] = None,  # noqa: ANN401
         raise_for_status: bool = True,
-    ) -> Tuple[Any, requests.Response]:
+    ) -> tuple[Any, requests.Response]:
         """
         Annofab WebAPIにアクセスして、レスポンスの中身とレスポンスを取得します。
 
@@ -484,7 +485,7 @@ class AnnofabApi(AbstractAnnofabApi):
             raise_for_status: Trueの場合HTTP Status Codeが4XX,5XXのときはHTTPErrorをスローします。Falseの場合はtuple[None, Response]を返します。
 
         Returns:
-            Tuple[content, Response]. contentはcontent_typeにより型が変わる。
+            tuple[content, Response]. contentはcontent_typeにより型が変わる。
             application/jsonならDict型, text/*ならばstr型, それ以外ならばbite型。
 
         Raises:
@@ -574,7 +575,7 @@ class AnnofabApi(AbstractAnnofabApi):
 
         return content, response
 
-    def _get_signed_cookie(self, project_id, query_params: Optional[Dict[str, Any]] = None) -> Tuple[Dict[str, Any], requests.Response]:  # noqa: ANN001
+    def _get_signed_cookie(self, project_id, query_params: Optional[dict[str, Any]] = None) -> tuple[dict[str, Any], requests.Response]:  # noqa: ANN001
         """
         アノテーション仕様の履歴情報を取得するために、非公開APIにアクセスする。
         変更される可能性あり.
@@ -583,12 +584,12 @@ class AnnofabApi(AbstractAnnofabApi):
             project_id: プロジェクトID
 
         Returns:
-            Tuple[Content, Response)
+            tuple[Content, Response)
 
         """
         url_path = f"/internal/projects/{project_id}/sign-headers"
         http_method = "GET"
-        keyword_params: Dict[str, Any] = {"query_params": query_params}
+        keyword_params: dict[str, Any] = {"query_params": query_params}
         return self._request_wrapper(http_method, url_path, **keyword_params)
 
     def _request_get_with_cookie(self, project_id: str, url: str) -> requests.Response:
@@ -625,7 +626,7 @@ class AnnofabApi(AbstractAnnofabApi):
     #########################################
     # Public Method : Login
     #########################################
-    def _login_respond_to_auth_challenge(self, id_pass: IdPass, mfa_code: str, session: str) -> Dict[str, Any]:
+    def _login_respond_to_auth_challenge(self, id_pass: IdPass, mfa_code: str, session: str) -> dict[str, Any]:
         """
         MFAコードによるログインを実行します。
 
@@ -674,7 +675,7 @@ class AnnofabApi(AbstractAnnofabApi):
             mfa_code: ``loginRespondToAuthChallenge``のレスポンスから取得したMFAコード。この引数はexperimentalです。将来削除される可能性があります。
 
         Returns:
-            Tuple[Token, requests.Response]
+            tuple[Token, requests.Response]
 
         Raises:
             InvalidMfaCodeError: ``self.input_mfa_code_via_stdin`` が ``False`` AND ``mfa_code`` が正しくない場合
@@ -722,7 +723,7 @@ class AnnofabApi(AbstractAnnofabApi):
 
 
         Returns:
-            Tuple[Token, requests.Response]
+            tuple[Token, requests.Response]
 
         Raises:
             NotLoggedInError: ログインしてない状態で関数を呼び出したときのエラー
