@@ -10,7 +10,7 @@ Note:
 """
 
 from dataclasses import dataclass
-from typing import Any, Dict  # pylint: disable=unused-import
+from typing import Any  # pylint: disable=unused-import
 
 from dataclasses_json import DataClassJsonMixin
 
@@ -33,7 +33,7 @@ class Organization(DataClassJsonMixin):
     price_plan: PricePlan
     """"""
 
-    summary: Dict[str, Any]
+    summary: dict[str, Any]
     """廃止予定のプロパティです。常に中身は空です。 """
 
     created_datetime: str
