@@ -9,7 +9,7 @@ Note:
 """
 
 import abc
-from typing import Any, Dict, Optional, Tuple  # pylint: disable=unused-import
+from typing import Any, Optional  # pylint: disable=unused-import
 
 import requests
 
@@ -24,10 +24,10 @@ class AbstractAnnofabApi2(abc.ABC):
         self,
         http_method: str,
         url_path: str,
-        query_params: Optional[Dict[str, Any]] = None,
-        header_params: Optional[Dict[str, Any]] = None,
+        query_params: Optional[dict[str, Any]] = None,
+        header_params: Optional[dict[str, Any]] = None,
         request_body: Optional[Any] = None,
-    ) -> Tuple[Any, requests.Response]:
+    ) -> tuple[Any, requests.Response]:
         pass
 
     #########################################
@@ -35,7 +35,7 @@ class AbstractAnnofabApi2(abc.ABC):
     # NOTE: This method is auto generated by OpenAPI Generator
     #########################################
 
-    def get_annotation_specs_v2(self, project_id: str, query_params: Optional[Dict[str, Any]] = None, **kwargs) -> Tuple[Any, requests.Response]:
+    def get_annotation_specs_v2(self, project_id: str, query_params: Optional[dict[str, Any]] = None, **kwargs) -> tuple[Any, requests.Response]:
         """アノテーション仕様取得
         https://annofab.com/docs/api/#operation/getAnnotationSpecsV2
 
@@ -47,19 +47,19 @@ class AbstractAnnofabApi2(abc.ABC):
 
         Args:
             project_id (str):  プロジェクトID (required)
-            query_params (Dict[str, Any]): Query Parameters
+            query_params (dict[str, Any]): Query Parameters
                 cache (str):  CACHE TIMESTAMP
                 history_id (str):  過去のアノテーション仕様を取得する場合、[アノテーション仕様履歴取得](#operation/getAnnotationSpecsHistories)APIで取得した `history_id` の値を指定します。 未指定時は最新のアノテーション仕様を取得します。
                 v (str):  取得するアノテーション仕様のフォーマットバージョンを指定します。
 
         Returns:
-            Tuple[AnnotationSpecs, requests.Response]
+            tuple[AnnotationSpecs, requests.Response]
 
 
         """
         url_path = f"/projects/{project_id}/annotation-specs"
         http_method = "GET"
-        keyword_params: Dict[str, Any] = {
+        keyword_params: dict[str, Any] = {
             "query_params": query_params,
         }
         keyword_params.update(**kwargs)
@@ -71,8 +71,8 @@ class AbstractAnnofabApi2(abc.ABC):
     #########################################
 
     def get_organization_member_v2(
-        self, organization_id: str, user_id: str, query_params: Optional[Dict[str, Any]] = None, **kwargs
-    ) -> Tuple[Any, requests.Response]:
+        self, organization_id: str, user_id: str, query_params: Optional[dict[str, Any]] = None, **kwargs
+    ) -> tuple[Any, requests.Response]:
         """組織メンバー取得
         https://annofab.com/docs/api/#operation/getOrganizationMemberV2
 
@@ -85,25 +85,25 @@ class AbstractAnnofabApi2(abc.ABC):
         Args:
             organization_id (str):  組織ID (required)
             user_id (str):  ユーザーID (required)
-            query_params (Dict[str, Any]): Query Parameters
+            query_params (dict[str, Any]): Query Parameters
                 cache (str):  CACHE TIMESTAMP
 
         Returns:
-            Tuple[OrganizationMember, requests.Response]
+            tuple[OrganizationMember, requests.Response]
 
 
         """
         url_path = f"/organizations/{organization_id}/members/{user_id}"
         http_method = "GET"
-        keyword_params: Dict[str, Any] = {
+        keyword_params: dict[str, Any] = {
             "query_params": query_params,
         }
         keyword_params.update(**kwargs)
         return self._request_wrapper(http_method, url_path, **keyword_params)
 
     def get_organization_members_v2(
-        self, organization_id: str, query_params: Optional[Dict[str, Any]] = None, **kwargs
-    ) -> Tuple[Any, requests.Response]:
+        self, organization_id: str, query_params: Optional[dict[str, Any]] = None, **kwargs
+    ) -> tuple[Any, requests.Response]:
         """組織メンバー一括取得
         https://annofab.com/docs/api/#operation/getOrganizationMembersV2
 
@@ -115,17 +115,17 @@ class AbstractAnnofabApi2(abc.ABC):
 
         Args:
             organization_id (str):  組織ID (required)
-            query_params (Dict[str, Any]): Query Parameters
+            query_params (dict[str, Any]): Query Parameters
                 cache (str):  CACHE TIMESTAMP
 
         Returns:
-            Tuple[OrganizationMemberList, requests.Response]
+            tuple[OrganizationMemberList, requests.Response]
 
 
         """
         url_path = f"/organizations/{organization_id}/members"
         http_method = "GET"
-        keyword_params: Dict[str, Any] = {
+        keyword_params: dict[str, Any] = {
             "query_params": query_params,
         }
         keyword_params.update(**kwargs)
@@ -137,8 +137,8 @@ class AbstractAnnofabApi2(abc.ABC):
     #########################################
 
     def get_organization_by_name_v2(
-        self, organization_name: str, query_params: Optional[Dict[str, Any]] = None, **kwargs
-    ) -> Tuple[Any, requests.Response]:
+        self, organization_name: str, query_params: Optional[dict[str, Any]] = None, **kwargs
+    ) -> tuple[Any, requests.Response]:
         """組織名で組織情報取得
         https://annofab.com/docs/api/#operation/getOrganizationByNameV2
 
@@ -150,23 +150,23 @@ class AbstractAnnofabApi2(abc.ABC):
 
         Args:
             organization_name (str):  組織名 (required)
-            query_params (Dict[str, Any]): Query Parameters
+            query_params (dict[str, Any]): Query Parameters
                 cache (str):  CACHE TIMESTAMP
 
         Returns:
-            Tuple[Organization, requests.Response]
+            tuple[Organization, requests.Response]
 
 
         """
         url_path = f"/organizations-by-name/{organization_name}"
         http_method = "GET"
-        keyword_params: Dict[str, Any] = {
+        keyword_params: dict[str, Any] = {
             "query_params": query_params,
         }
         keyword_params.update(**kwargs)
         return self._request_wrapper(http_method, url_path, **keyword_params)
 
-    def get_organization_cache_by_name_v2(self, organization_name: str, **kwargs) -> Tuple[Any, requests.Response]:
+    def get_organization_cache_by_name_v2(self, organization_name: str, **kwargs) -> tuple[Any, requests.Response]:
         """組織名で組織キャッシュレコード取得
         https://annofab.com/docs/api/#operation/getOrganizationCacheByNameV2
 
@@ -180,17 +180,17 @@ class AbstractAnnofabApi2(abc.ABC):
             organization_name (str):  組織名 (required)
 
         Returns:
-            Tuple[OrganizationCacheRecord, requests.Response]
+            tuple[OrganizationCacheRecord, requests.Response]
 
 
         """
         url_path = f"/organizations-by-name/{organization_name}/cache"
         http_method = "GET"
-        keyword_params: Dict[str, Any] = {}
+        keyword_params: dict[str, Any] = {}
         keyword_params.update(**kwargs)
         return self._request_wrapper(http_method, url_path, **keyword_params)
 
-    def get_organization_cache_v2(self, organization_id: str, **kwargs) -> Tuple[Any, requests.Response]:
+    def get_organization_cache_v2(self, organization_id: str, **kwargs) -> tuple[Any, requests.Response]:
         """組織IDで組織キャッシュレコード取得
         https://annofab.com/docs/api/#operation/getOrganizationCacheV2
 
@@ -204,17 +204,17 @@ class AbstractAnnofabApi2(abc.ABC):
             organization_id (str):  組織ID (required)
 
         Returns:
-            Tuple[OrganizationCacheRecord, requests.Response]
+            tuple[OrganizationCacheRecord, requests.Response]
 
 
         """
         url_path = f"/organizations/{organization_id}/cache"
         http_method = "GET"
-        keyword_params: Dict[str, Any] = {}
+        keyword_params: dict[str, Any] = {}
         keyword_params.update(**kwargs)
         return self._request_wrapper(http_method, url_path, **keyword_params)
 
-    def get_organization_v2(self, organization_id: str, query_params: Optional[Dict[str, Any]] = None, **kwargs) -> Tuple[Any, requests.Response]:
+    def get_organization_v2(self, organization_id: str, query_params: Optional[dict[str, Any]] = None, **kwargs) -> tuple[Any, requests.Response]:
         """組織IDで組織情報取得
         https://annofab.com/docs/api/#operation/getOrganizationV2
 
@@ -226,23 +226,23 @@ class AbstractAnnofabApi2(abc.ABC):
 
         Args:
             organization_id (str):  組織ID (required)
-            query_params (Dict[str, Any]): Query Parameters
+            query_params (dict[str, Any]): Query Parameters
                 cache (str):  CACHE TIMESTAMP
 
         Returns:
-            Tuple[Organization, requests.Response]
+            tuple[Organization, requests.Response]
 
 
         """
         url_path = f"/organizations/{organization_id}"
         http_method = "GET"
-        keyword_params: Dict[str, Any] = {
+        keyword_params: dict[str, Any] = {
             "query_params": query_params,
         }
         keyword_params.update(**kwargs)
         return self._request_wrapper(http_method, url_path, **keyword_params)
 
-    def get_project_task_count_v2(self, organization_id: str, project_id: str, **kwargs) -> Tuple[Any, requests.Response]:
+    def get_project_task_count_v2(self, organization_id: str, project_id: str, **kwargs) -> tuple[Any, requests.Response]:
         """プロジェクトのタスク件数取得
         https://annofab.com/docs/api/#operation/getProjectTaskCountV2
 
@@ -257,13 +257,13 @@ class AbstractAnnofabApi2(abc.ABC):
             project_id (str):  プロジェクトID (required)
 
         Returns:
-            Tuple[ProjectTaskCounts, requests.Response]
+            tuple[ProjectTaskCounts, requests.Response]
 
 
         """
         url_path = f"/organizations/{organization_id}/projects/{project_id}/task-counts"
         http_method = "GET"
-        keyword_params: Dict[str, Any] = {}
+        keyword_params: dict[str, Any] = {}
         keyword_params.update(**kwargs)
         return self._request_wrapper(http_method, url_path, **keyword_params)
 
@@ -273,8 +273,8 @@ class AbstractAnnofabApi2(abc.ABC):
     #########################################
 
     def get_project_member_v2(
-        self, project_id: str, user_id: str, query_params: Optional[Dict[str, Any]] = None, **kwargs
-    ) -> Tuple[Any, requests.Response]:
+        self, project_id: str, user_id: str, query_params: Optional[dict[str, Any]] = None, **kwargs
+    ) -> tuple[Any, requests.Response]:
         """プロジェクトメンバー取得
         https://annofab.com/docs/api/#operation/getProjectMemberV2
 
@@ -287,23 +287,23 @@ class AbstractAnnofabApi2(abc.ABC):
         Args:
             project_id (str):  プロジェクトID (required)
             user_id (str):  アカウントのユーザーID. RESTクライアントユーザーが指定しやすいように、Cognitoのaccount_idではなくuser_idとしている。 (required)
-            query_params (Dict[str, Any]): Query Parameters
+            query_params (dict[str, Any]): Query Parameters
                 cache (str):  CACHE TIMESTAMP
 
         Returns:
-            Tuple[ProjectMember, requests.Response]
+            tuple[ProjectMember, requests.Response]
 
 
         """
         url_path = f"/projects/{project_id}/members/{user_id}"
         http_method = "GET"
-        keyword_params: Dict[str, Any] = {
+        keyword_params: dict[str, Any] = {
             "query_params": query_params,
         }
         keyword_params.update(**kwargs)
         return self._request_wrapper(http_method, url_path, **keyword_params)
 
-    def get_project_members_v2(self, project_id: str, query_params: Optional[Dict[str, Any]] = None, **kwargs) -> Tuple[Any, requests.Response]:
+    def get_project_members_v2(self, project_id: str, query_params: Optional[dict[str, Any]] = None, **kwargs) -> tuple[Any, requests.Response]:
         """プロジェクトメンバー検索
         https://annofab.com/docs/api/#operation/getProjectMembersV2
 
@@ -315,18 +315,18 @@ class AbstractAnnofabApi2(abc.ABC):
 
         Args:
             project_id (str):  プロジェクトID (required)
-            query_params (Dict[str, Any]): Query Parameters
+            query_params (dict[str, Any]): Query Parameters
                 include_inactive_member (str):  脱退したプロジェクトメンバーも取得する時に、キーのみ指定します（値は無視されます）。
                 cache (str):  CACHE TIMESTAMP
 
         Returns:
-            Tuple[ProjectMemberList, requests.Response]
+            tuple[ProjectMemberList, requests.Response]
 
 
         """
         url_path = f"/projects/{project_id}/members"
         http_method = "GET"
-        keyword_params: Dict[str, Any] = {
+        keyword_params: dict[str, Any] = {
             "query_params": query_params,
         }
         keyword_params.update(**kwargs)
@@ -337,7 +337,7 @@ class AbstractAnnofabApi2(abc.ABC):
     # NOTE: This method is auto generated by OpenAPI Generator
     #########################################
 
-    def get_project_cache_v2(self, project_id: str, **kwargs) -> Tuple[Any, requests.Response]:
+    def get_project_cache_v2(self, project_id: str, **kwargs) -> tuple[Any, requests.Response]:
         """プロジェクトキャッシュレコード取得
         https://annofab.com/docs/api/#operation/getProjectCacheV2
 
@@ -351,13 +351,13 @@ class AbstractAnnofabApi2(abc.ABC):
             project_id (str):  プロジェクトID (required)
 
         Returns:
-            Tuple[ProjectCacheRecord, requests.Response]
+            tuple[ProjectCacheRecord, requests.Response]
 
 
         """
         url_path = f"/projects/{project_id}/cache"
         http_method = "GET"
-        keyword_params: Dict[str, Any] = {}
+        keyword_params: dict[str, Any] = {}
         keyword_params.update(**kwargs)
         return self._request_wrapper(http_method, url_path, **keyword_params)
 
@@ -367,8 +367,8 @@ class AbstractAnnofabApi2(abc.ABC):
     #########################################
 
     def get_account_daily_statistics_v2(
-        self, project_id: str, query_params: Optional[Dict[str, Any]] = None, **kwargs
-    ) -> Tuple[Any, requests.Response]:
+        self, project_id: str, query_params: Optional[dict[str, Any]] = None, **kwargs
+    ) -> tuple[Any, requests.Response]:
         """ユーザー別タスク集計取得
         https://annofab.com/docs/api/#operation/getAccountDailyStatisticsV2
 
@@ -380,26 +380,26 @@ class AbstractAnnofabApi2(abc.ABC):
 
         Args:
             project_id (str):  プロジェクトID (required)
-            query_params (Dict[str, Any]): Query Parameters
+            query_params (dict[str, Any]): Query Parameters
                 from (str):  取得する統計の区間の開始日 - `YYYY-MM-DD` (required)
                 to (str):  取得する統計の区間の終了日 - `YYYY-MM-DD` 。この日は含まれます。 (required)
 
         Returns:
-            Tuple[List[ProjectAccountStatistics], requests.Response]
+            tuple[list[ProjectAccountStatistics], requests.Response]
 
 
         """
         url_path = f"/projects/{project_id}/statistics/accounts/daily"
         http_method = "GET"
-        keyword_params: Dict[str, Any] = {
+        keyword_params: dict[str, Any] = {
             "query_params": query_params,
         }
         keyword_params.update(**kwargs)
         return self._request_wrapper(http_method, url_path, **keyword_params)
 
     def get_inspection_daily_statistics_v2(
-        self, project_id: str, query_params: Optional[Dict[str, Any]] = None, **kwargs
-    ) -> Tuple[Any, requests.Response]:
+        self, project_id: str, query_params: Optional[dict[str, Any]] = None, **kwargs
+    ) -> tuple[Any, requests.Response]:
         """検査コメント集計取得
         https://annofab.com/docs/api/#operation/getInspectionDailyStatisticsV2
 
@@ -411,24 +411,24 @@ class AbstractAnnofabApi2(abc.ABC):
 
         Args:
             project_id (str):  プロジェクトID (required)
-            query_params (Dict[str, Any]): Query Parameters
+            query_params (dict[str, Any]): Query Parameters
                 from (str):  取得する統計の区間の開始日 - `YYYY-MM-DD` (required)
                 to (str):  取得する統計の区間の終了日 - `YYYY-MM-DD` 。この日は含まれます。 (required)
 
         Returns:
-            Tuple[List[InspectionStatistics], requests.Response]
+            tuple[list[InspectionStatistics], requests.Response]
 
 
         """
         url_path = f"/projects/{project_id}/statistics/inspections/daily"
         http_method = "GET"
-        keyword_params: Dict[str, Any] = {
+        keyword_params: dict[str, Any] = {
             "query_params": query_params,
         }
         keyword_params.update(**kwargs)
         return self._request_wrapper(http_method, url_path, **keyword_params)
 
-    def get_label_statistics_v2(self, project_id: str, query_params: Optional[Dict[str, Any]] = None, **kwargs) -> Tuple[Any, requests.Response]:
+    def get_label_statistics_v2(self, project_id: str, query_params: Optional[dict[str, Any]] = None, **kwargs) -> tuple[Any, requests.Response]:
         """ラベル別アノテーション数集計取得
         https://annofab.com/docs/api/#operation/getLabelStatisticsV2
 
@@ -440,25 +440,25 @@ class AbstractAnnofabApi2(abc.ABC):
 
         Args:
             project_id (str):  プロジェクトID (required)
-            query_params (Dict[str, Any]): Query Parameters
+            query_params (dict[str, Any]): Query Parameters
                 cache (str):  CACHE TIMESTAMP
 
         Returns:
-            Tuple[List[LabelStatistics], requests.Response]
+            tuple[list[LabelStatistics], requests.Response]
 
 
         """
         url_path = f"/projects/{project_id}/statistics/labels"
         http_method = "GET"
-        keyword_params: Dict[str, Any] = {
+        keyword_params: dict[str, Any] = {
             "query_params": query_params,
         }
         keyword_params.update(**kwargs)
         return self._request_wrapper(http_method, url_path, **keyword_params)
 
     def get_phase_daily_statistics_v2(
-        self, project_id: str, query_params: Optional[Dict[str, Any]] = None, **kwargs
-    ) -> Tuple[Any, requests.Response]:
+        self, project_id: str, query_params: Optional[dict[str, Any]] = None, **kwargs
+    ) -> tuple[Any, requests.Response]:
         """フェーズ別タスク集計取得
         https://annofab.com/docs/api/#operation/getPhaseDailyStatisticsV2
 
@@ -470,24 +470,24 @@ class AbstractAnnofabApi2(abc.ABC):
 
         Args:
             project_id (str):  プロジェクトID (required)
-            query_params (Dict[str, Any]): Query Parameters
+            query_params (dict[str, Any]): Query Parameters
                 from (str):  取得する統計の区間の開始日 - `YYYY-MM-DD` (required)
                 to (str):  取得する統計の区間の終了日 - `YYYY-MM-DD` 。この日は含まれます。 (required)
 
         Returns:
-            Tuple[List[TaskPhaseStatistics], requests.Response]
+            tuple[list[TaskPhaseStatistics], requests.Response]
 
 
         """
         url_path = f"/projects/{project_id}/statistics/task-phases/daily"
         http_method = "GET"
-        keyword_params: Dict[str, Any] = {
+        keyword_params: dict[str, Any] = {
             "query_params": query_params,
         }
         keyword_params.update(**kwargs)
         return self._request_wrapper(http_method, url_path, **keyword_params)
 
-    def get_task_daily_statistics_v2(self, project_id: str, query_params: Optional[Dict[str, Any]] = None, **kwargs) -> Tuple[Any, requests.Response]:
+    def get_task_daily_statistics_v2(self, project_id: str, query_params: Optional[dict[str, Any]] = None, **kwargs) -> tuple[Any, requests.Response]:
         """タスク集計取得
         https://annofab.com/docs/api/#operation/getTaskDailyStatisticsV2
 
@@ -499,26 +499,26 @@ class AbstractAnnofabApi2(abc.ABC):
 
         Args:
             project_id (str):  プロジェクトID (required)
-            query_params (Dict[str, Any]): Query Parameters
+            query_params (dict[str, Any]): Query Parameters
                 from (str):  取得する統計の区間の開始日 - `YYYY-MM-DD` (required)
                 to (str):  取得する統計の区間の終了日 - `YYYY-MM-DD` 。この日は含まれます。 (required)
 
         Returns:
-            Tuple[List[ProjectTaskStatisticsHistory], requests.Response]
+            tuple[list[ProjectTaskStatisticsHistory], requests.Response]
 
 
         """
         url_path = f"/projects/{project_id}/statistics/tasks/daily"
         http_method = "GET"
-        keyword_params: Dict[str, Any] = {
+        keyword_params: dict[str, Any] = {
             "query_params": query_params,
         }
         keyword_params.update(**kwargs)
         return self._request_wrapper(http_method, url_path, **keyword_params)
 
     def get_worktime_daily_statistics_by_account_v2(
-        self, project_id: str, account_id: str, query_params: Optional[Dict[str, Any]] = None, **kwargs
-    ) -> Tuple[Any, requests.Response]:
+        self, project_id: str, account_id: str, query_params: Optional[dict[str, Any]] = None, **kwargs
+    ) -> tuple[Any, requests.Response]:
         """プロジェクトメンバー単位のフェーズ別タスク集計取得
         https://annofab.com/docs/api/#operation/getWorktimeDailyStatisticsByAccountV2
 
@@ -531,26 +531,26 @@ class AbstractAnnofabApi2(abc.ABC):
         Args:
             project_id (str):  プロジェクトID (required)
             account_id (str):  アカウントID (required)
-            query_params (Dict[str, Any]): Query Parameters
+            query_params (dict[str, Any]): Query Parameters
                 from (str):  取得する統計の区間の開始日 - `YYYY-MM-DD` (required)
                 to (str):  取得する統計の区間の終了日 - `YYYY-MM-DD` 。この日は含まれます。 (required)
 
         Returns:
-            Tuple[WorktimeStatisticsByAccount, requests.Response]
+            tuple[WorktimeStatisticsByAccount, requests.Response]
 
 
         """
         url_path = f"/projects/{project_id}/statistics/worktimes-by-account/{account_id}/daily"
         http_method = "GET"
-        keyword_params: Dict[str, Any] = {
+        keyword_params: dict[str, Any] = {
             "query_params": query_params,
         }
         keyword_params.update(**kwargs)
         return self._request_wrapper(http_method, url_path, **keyword_params)
 
     def get_worktime_daily_statistics_v2(
-        self, project_id: str, query_params: Optional[Dict[str, Any]] = None, **kwargs
-    ) -> Tuple[Any, requests.Response]:
+        self, project_id: str, query_params: Optional[dict[str, Any]] = None, **kwargs
+    ) -> tuple[Any, requests.Response]:
         """プロジェクト単位のフェーズ別タスク集計取得
         https://annofab.com/docs/api/#operation/getWorktimeDailyStatisticsV2
 
@@ -562,18 +562,18 @@ class AbstractAnnofabApi2(abc.ABC):
 
         Args:
             project_id (str):  プロジェクトID (required)
-            query_params (Dict[str, Any]): Query Parameters
+            query_params (dict[str, Any]): Query Parameters
                 from (str):  取得する統計の区間の開始日 - `YYYY-MM-DD` (required)
                 to (str):  取得する統計の区間の終了日 - `YYYY-MM-DD` 。この日は含まれます。 (required)
 
         Returns:
-            Tuple[WorktimeStatisticsByProject, requests.Response]
+            tuple[WorktimeStatisticsByProject, requests.Response]
 
 
         """
         url_path = f"/projects/{project_id}/statistics/worktimes/daily"
         http_method = "GET"
-        keyword_params: Dict[str, Any] = {
+        keyword_params: dict[str, Any] = {
             "query_params": query_params,
         }
         keyword_params.update(**kwargs)
@@ -584,7 +584,7 @@ class AbstractAnnofabApi2(abc.ABC):
     # NOTE: This method is auto generated by OpenAPI Generator
     #########################################
 
-    def get_account_v2(self, account_id: str, query_params: Optional[Dict[str, Any]] = None, **kwargs) -> Tuple[Any, requests.Response]:
+    def get_account_v2(self, account_id: str, query_params: Optional[dict[str, Any]] = None, **kwargs) -> tuple[Any, requests.Response]:
         """ユーザーアカウント情報取得
         https://annofab.com/docs/api/#operation/getAccountV2
 
@@ -596,23 +596,23 @@ class AbstractAnnofabApi2(abc.ABC):
 
         Args:
             account_id (str):  アカウントID (required)
-            query_params (Dict[str, Any]): Query Parameters
+            query_params (dict[str, Any]): Query Parameters
                 cache (str):  CACHE TIMESTAMP
 
         Returns:
-            Tuple[MyAccount, requests.Response]
+            tuple[MyAccount, requests.Response]
 
 
         """
         url_path = f"/users/{account_id}"
         http_method = "GET"
-        keyword_params: Dict[str, Any] = {
+        keyword_params: dict[str, Any] = {
             "query_params": query_params,
         }
         keyword_params.update(**kwargs)
         return self._request_wrapper(http_method, url_path, **keyword_params)
 
-    def get_user_cache_v2(self, account_id: str, **kwargs) -> Tuple[Any, requests.Response]:
+    def get_user_cache_v2(self, account_id: str, **kwargs) -> tuple[Any, requests.Response]:
         """キャッシュレコード
         https://annofab.com/docs/api/#operation/getUserCacheV2
 
@@ -626,17 +626,17 @@ class AbstractAnnofabApi2(abc.ABC):
             account_id (str):  アカウントID (required)
 
         Returns:
-            Tuple[UserCacheRecord, requests.Response]
+            tuple[UserCacheRecord, requests.Response]
 
 
         """
         url_path = f"/users/{account_id}/cache"
         http_method = "GET"
-        keyword_params: Dict[str, Any] = {}
+        keyword_params: dict[str, Any] = {}
         keyword_params.update(**kwargs)
         return self._request_wrapper(http_method, url_path, **keyword_params)
 
-    def get_user_organizations_v2(self, account_id: str, query_params: Optional[Dict[str, Any]] = None, **kwargs) -> Tuple[Any, requests.Response]:
+    def get_user_organizations_v2(self, account_id: str, query_params: Optional[dict[str, Any]] = None, **kwargs) -> tuple[Any, requests.Response]:
         """個人の所属組織検索
         https://annofab.com/docs/api/#operation/getUserOrganizationsV2
 
@@ -648,23 +648,23 @@ class AbstractAnnofabApi2(abc.ABC):
 
         Args:
             account_id (str):  アカウントID (required)
-            query_params (Dict[str, Any]): Query Parameters
+            query_params (dict[str, Any]): Query Parameters
                 cache (str):  CACHE TIMESTAMP
 
         Returns:
-            Tuple[MyOrganizationList, requests.Response]
+            tuple[MyOrganizationList, requests.Response]
 
 
         """
         url_path = f"/users/{account_id}/organizations"
         http_method = "GET"
-        keyword_params: Dict[str, Any] = {
+        keyword_params: dict[str, Any] = {
             "query_params": query_params,
         }
         keyword_params.update(**kwargs)
         return self._request_wrapper(http_method, url_path, **keyword_params)
 
-    def get_user_project_members_v2(self, account_id: str, query_params: Optional[Dict[str, Any]] = None, **kwargs) -> Tuple[Any, requests.Response]:
+    def get_user_project_members_v2(self, account_id: str, query_params: Optional[dict[str, Any]] = None, **kwargs) -> tuple[Any, requests.Response]:
         """個人のプロジェクトメンバー情報一括取得
         https://annofab.com/docs/api/#operation/getUserProjectMembersV2
 
@@ -676,23 +676,23 @@ class AbstractAnnofabApi2(abc.ABC):
 
         Args:
             account_id (str):  アカウントID (required)
-            query_params (Dict[str, Any]): Query Parameters
+            query_params (dict[str, Any]): Query Parameters
                 cache (str):  CACHE TIMESTAMP
 
         Returns:
-            Tuple[List[ProjectMember], requests.Response]
+            tuple[list[ProjectMember], requests.Response]
 
 
         """
         url_path = f"/users/{account_id}/project-members"
         http_method = "GET"
-        keyword_params: Dict[str, Any] = {
+        keyword_params: dict[str, Any] = {
             "query_params": query_params,
         }
         keyword_params.update(**kwargs)
         return self._request_wrapper(http_method, url_path, **keyword_params)
 
-    def get_user_projects_v2(self, account_id: str, query_params: Optional[Dict[str, Any]] = None, **kwargs) -> Tuple[Any, requests.Response]:
+    def get_user_projects_v2(self, account_id: str, query_params: Optional[dict[str, Any]] = None, **kwargs) -> tuple[Any, requests.Response]:
         """自分のプロジェクトメンバー情報検索
         https://annofab.com/docs/api/#operation/getUserProjectsV2
 
@@ -704,7 +704,7 @@ class AbstractAnnofabApi2(abc.ABC):
 
         Args:
             account_id (str):  アカウントID (required)
-            query_params (Dict[str, Any]): Query Parameters
+            query_params (dict[str, Any]): Query Parameters
                 page (int):  表示するページ番号
                 limit (int):  1ページあたりの取得するデータ件数
                 organization_id (str):  指定した組織に属するプロジェクトに絞り込む。未指定時は全プロジェクト。
@@ -715,13 +715,13 @@ class AbstractAnnofabApi2(abc.ABC):
                 cache (str):  CACHE TIMESTAMP
 
         Returns:
-            Tuple[ProjectContainer, requests.Response]
+            tuple[ProjectContainer, requests.Response]
 
 
         """
         url_path = f"/users/{account_id}/projects"
         http_method = "GET"
-        keyword_params: Dict[str, Any] = {
+        keyword_params: dict[str, Any] = {
             "query_params": query_params,
         }
         keyword_params.update(**kwargs)
