@@ -11,7 +11,7 @@ Note:
 import abc
 import annofabapi  # pylint: disable=unused-import
 import warnings  # pylint: disable=unused-import
-from typing import Any, Dict, List, Optional, Tuple, Union  # pylint: disable=unused-import
+from typing import Any, Optional, Union  # pylint: disable=unused-import
 import requests
 
 
@@ -25,8 +25,8 @@ class AbstractAnnofabApi(abc.ABC):
         self,
         http_method: str,
         url_path: str,
-        query_params: Optional[Dict[str, Any]] = None,
-        header_params: Optional[Dict[str, Any]] = None,
+        query_params: Optional[dict[str, Any]] = None,
+        header_params: Optional[dict[str, Any]] = None,
         request_body: Optional[Any] = None,
-    ) -> Tuple[Any, requests.Response]:
+    ) -> tuple[Any, requests.Response]:
         pass
