@@ -12,7 +12,7 @@ Note:
 
 import warnings  # pylint: disable=unused-import
 from enum import Enum
-from typing import Any, Dict, List, NewType, Optional, Tuple, Union  # pylint: disable=unused-import
+from typing import Any, NewType, Optional, Union  # pylint: disable=unused-import
 
 from annofabapi._utils import deprecated_class  # pylint: disable=unused-import
 from annofabapi._utils import deprecated_class  # pylint: disable=unused-import
